Move test_st_arrow_table to arrow_dataframe_test.py file
