+++ lib/tests/streamlit/elements/arrow_dataframe_test.py
@@ -160,3 +160,18 @@ class ArrowDataFrameProtoTest(DeltaGeneratorTestCase):
 
         proto = self.get_delta_from_queue().new_element.arrow_data_frame
         pd.testing.assert_frame_equal(bytes_to_data_frame(proto.data), expected)
+
+
+class StArrowTableAPITest(DeltaGeneratorTestCase):
+    """Test Public Streamlit Public APIs."""
+
+    def test_st_arrow_table(self):
+        """Test st._arrow_table."""
+        from streamlit.type_util import bytes_to_data_frame
+
+        df = pd.DataFrame([[1, 2], [3, 4]], columns=["col1", "col2"])
+
+        st._arrow_table(df)
+
+        proto = self.get_delta_from_queue().new_element.arrow_table
+        pd.testing.assert_frame_equal(bytes_to_data_frame(proto.data), df)
